requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jpype1==1.5.2
MarkupSafe==3.0.2
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pycryptodome==3.22.0
PyMySQL==1.1.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
tabula-py==2.10.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
XlsxWriter==3.2.3
